+++ func/uk/gov/hmrc/r2/selfassessmentapi/resources/PropertiesAnnualSummaryResourceISpec.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.selfassessmentapi.resources
+package uk.gov.hmrc.r2.selfassessmentapi.resources
 
-import uk.gov.hmrc.selfassessmentapi.models.properties.PropertyType
-import uk.gov.hmrc.support.BaseFunctionalSpec
+import uk.gov.hmrc.r2.selfassessmentapi.models.properties.PropertyType
+import uk.gov.hmrc.r2.selfassessmentapi.support.BaseFunctionalSpec
 
-class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
+class PropertiesAnnualSummaryResourceISpec extends BaseFunctionalSpec {
 
   "amending annual summaries" should {
     for (propertyType <- Seq(PropertyType.OTHER, PropertyType.FHL)) {
@@ -29,7 +29,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillBeUpdatedFor(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(204)
       }
@@ -43,7 +43,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .put(invalidAnnualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(invalidAnnualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(400)
           .contentTypeIsJson()
@@ -54,9 +54,12 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
         given()
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
+          .des()
+          .properties
+          .willBeCreatedFor(nino)
           .des().properties.annualSummaryWillNotBeReturnedFor(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -67,7 +70,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundProperty(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -78,7 +81,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().invalidOriginatorIdFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .bodyIsLike(Jsons.Errors.internalServerError)
@@ -90,7 +93,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().payloadFailsValidationFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(400)
           .contentTypeIsJson()
@@ -102,22 +105,10 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/2015-16")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/2015-16")
           .thenAssertThat()
           .statusIs(400)
           .bodyIsError("TAX_YEAR_INVALID")
-      }
-
-      s"return code 404 when attempting to update annual summaries for an invalid property type for $propertyType" in {
-        given()
-          .userIsSubscribedToMtdFor(nino)
-          .clientIsFullyAuthorisedForTheResource
-          .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/silly/$taxYear")
-          .thenAssertThat()
-          .statusIs(404)
-          .contentTypeIsJson()
-          .bodyIsLike(Jsons.Errors.notFound)
       }
 
       s"return code 500 when DES is experiencing problems for $propertyType" in {
@@ -126,7 +117,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serverErrorFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -139,7 +130,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serviceUnavailableFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -152,7 +143,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().isATeapotFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
       }
@@ -166,7 +157,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
         .userIsSubscribedToMtdFor(nino)
         .clientIsFullyAuthorisedForTheResource
         .when()
-        .put(s"/ni/$nino/uk-properties/${PropertyType.FHL}/$taxYear", Some(Jsons.Properties.invalidFhlAnnualSummary))
+        .put(s"/r2/ni/$nino/uk-properties/${PropertyType.FHL}/$taxYear", Some(Jsons.Properties.invalidFhlAnnualSummary))
         .thenAssertThat()
         .statusIs(400)
         .contentTypeIsJson()
@@ -183,7 +174,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillBeReturnedFor(nino, propertyType, taxYear, desAnnualSummary(propertyType))
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(200)
           .contentTypeIsJson()
@@ -196,7 +187,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.noAnnualSummaryFor(nino, propertyType, taxYear)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -207,7 +198,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundProperty(nino, propertyType, taxYear)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -218,7 +209,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundPeriod(nino, propertyType, taxYear)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -228,7 +219,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/2015-16")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/2015-16")
           .thenAssertThat()
           .statusIs(400)
           .bodyIsError("TAX_YEAR_INVALID")
@@ -240,7 +231,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().invalidOriginatorIdFor(nino)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .bodyIsLike(Jsons.Errors.internalServerError)
@@ -252,7 +243,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serverErrorFor(nino)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -265,7 +256,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serviceUnavailableFor(nino)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -278,7 +269,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().isATeapotFor(nino)
           .when()
-          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
       }
@@ -286,7 +277,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
   }
 
   private def annualSummary(propertyType: PropertyType.Value) = propertyType match {
-    case PropertyType.OTHER => Jsons.Properties.otherAnnualSummary()
+    case PropertyType.OTHER => Jsons.Properties.otherAnnualSummary(rarJointLet = false)
     case PropertyType.FHL => Jsons.Properties.fhlAnnualSummary()
   }
 
@@ -298,7 +289,11 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
       costOfReplacingDomesticItems = 150.55,
       lossBroughtForward = 20.22,
       privateUseAdjustment = -22.23,
-      balancingCharge = 350.34)
+      balancingCharge = 350.34,
+      bpraBalancingCharge = 0.0,
+      nonResidentLandlord = true,
+      rarJointLet = false
+    )
     case PropertyType.FHL => Jsons.Properties.fhlAnnualSummary(
       annualInvestmentAllowance = -10000.50,
       otherCapitalAllowance = 1000.20,
