+++ func/uk/gov/hmrc/selfassessmentapi/resources/PropertiesAnnualSummaryResourceISpec.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.r2.selfassessmentapi.resources
+package uk.gov.hmrc.selfassessmentapi.resources
 
-import uk.gov.hmrc.r2.selfassessmentapi.models.properties.PropertyType
-import uk.gov.hmrc.r2.selfassessmentapi.support.BaseFunctionalSpec
+import uk.gov.hmrc.selfassessmentapi.models.properties.PropertyType
+import uk.gov.hmrc.support.BaseFunctionalSpec
 
-class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
+class PropertiesAnnualSummaryResourceISpec extends BaseFunctionalSpec {
 
   "amending annual summaries" should {
     for (propertyType <- Seq(PropertyType.OTHER, PropertyType.FHL)) {
@@ -29,7 +29,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillBeUpdatedFor(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(204)
       }
@@ -43,7 +43,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .put(invalidAnnualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(invalidAnnualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(400)
           .contentTypeIsJson()
@@ -54,12 +54,9 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
         given()
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
-          .des()
-          .properties
-          .willBeCreatedFor(nino)
           .des().properties.annualSummaryWillNotBeReturnedFor(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -70,7 +67,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundProperty(nino, propertyType, taxYear)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -81,7 +78,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().invalidOriginatorIdFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .bodyIsLike(Jsons.Errors.internalServerError)
@@ -93,7 +90,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().payloadFailsValidationFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(400)
           .contentTypeIsJson()
@@ -105,10 +102,22 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/2015-16")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/2015-16")
           .thenAssertThat()
           .statusIs(400)
           .bodyIsError("TAX_YEAR_INVALID")
+      }
+
+      s"return code 404 when attempting to update annual summaries for an invalid property type for $propertyType" in {
+        given()
+          .userIsSubscribedToMtdFor(nino)
+          .clientIsFullyAuthorisedForTheResource
+          .when()
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/silly/$taxYear")
+          .thenAssertThat()
+          .statusIs(404)
+          .contentTypeIsJson()
+          .bodyIsLike(Jsons.Errors.notFound)
       }
 
       s"return code 500 when DES is experiencing problems for $propertyType" in {
@@ -117,7 +126,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serverErrorFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -130,7 +139,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serviceUnavailableFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -143,7 +152,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().isATeapotFor(nino)
           .when()
-          .put(annualSummary(propertyType)).at(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .put(annualSummary(propertyType)).at(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
       }
@@ -157,7 +166,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
         .userIsSubscribedToMtdFor(nino)
         .clientIsFullyAuthorisedForTheResource
         .when()
-        .put(s"/r2/ni/$nino/uk-properties/${PropertyType.FHL}/$taxYear", Some(Jsons.Properties.invalidFhlAnnualSummary))
+        .put(s"/ni/$nino/uk-properties/${PropertyType.FHL}/$taxYear", Some(Jsons.Properties.invalidFhlAnnualSummary))
         .thenAssertThat()
         .statusIs(400)
         .contentTypeIsJson()
@@ -174,7 +183,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillBeReturnedFor(nino, propertyType, taxYear, desAnnualSummary(propertyType))
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(200)
           .contentTypeIsJson()
@@ -187,7 +196,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.noAnnualSummaryFor(nino, propertyType, taxYear)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -198,7 +207,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundProperty(nino, propertyType, taxYear)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -209,7 +218,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().properties.annualSummaryWillNotBeReturnedDueToNotFoundPeriod(nino, propertyType, taxYear)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(404)
       }
@@ -219,7 +228,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .userIsSubscribedToMtdFor(nino)
           .clientIsFullyAuthorisedForTheResource
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/2015-16")
+          .get(s"/ni/$nino/uk-properties/$propertyType/2015-16")
           .thenAssertThat()
           .statusIs(400)
           .bodyIsError("TAX_YEAR_INVALID")
@@ -231,7 +240,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().invalidOriginatorIdFor(nino)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .bodyIsLike(Jsons.Errors.internalServerError)
@@ -243,7 +252,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serverErrorFor(nino)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -256,7 +265,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().serviceUnavailableFor(nino)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
           .contentTypeIsJson()
@@ -269,7 +278,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
           .clientIsFullyAuthorisedForTheResource
           .des().isATeapotFor(nino)
           .when()
-          .get(s"/r2/ni/$nino/uk-properties/$propertyType/$taxYear")
+          .get(s"/ni/$nino/uk-properties/$propertyType/$taxYear")
           .thenAssertThat()
           .statusIs(500)
       }
@@ -277,7 +286,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
   }
 
   private def annualSummary(propertyType: PropertyType.Value) = propertyType match {
-    case PropertyType.OTHER => Jsons.Properties.otherAnnualSummary(rarJointLet = false)
+    case PropertyType.OTHER => Jsons.Properties.otherAnnualSummary()
     case PropertyType.FHL => Jsons.Properties.fhlAnnualSummary()
   }
 
@@ -289,11 +298,7 @@ class PropertiesAnnualSummaryResourceSpec extends BaseFunctionalSpec {
       costOfReplacingDomesticItems = 150.55,
       lossBroughtForward = 20.22,
       privateUseAdjustment = -22.23,
-      balancingCharge = 350.34,
-      bpraBalancingCharge = 0.0,
-      nonResidentLandlord = true,
-      rarJointLet = false
-    )
+      balancingCharge = 350.34)
     case PropertyType.FHL => Jsons.Properties.fhlAnnualSummary(
       annualInvestmentAllowance = -10000.50,
       otherCapitalAllowance = 1000.20,
